daprApps_v1/video-pipe/video-meta/requirements.txt
# dapr-ext-grpc-dev >= 1.8.0rc1.dev
# dapr-dev >= 1.8.0rc1.dev
dapr==1.8.3
dapr-ext-grpc==1.8.3
prometheus-client
ffmpeg-python
numpy